@@ -1,10 +1,11 @@
-import type { Prisma } from '@prisma/client';
-import { StatusCodes } from 'http-status-codes';
 import { getServerSession } from 'next-auth';
 import { NextResponse } from 'next/server';
+
+import type { Prisma } from '@prisma/client';
+import { StatusCodes } from 'http-status-codes';
 import { parse } from 'qs';
 
-import { authOptions, DEFAULT_PAGE_SIZE, FALSE, TRUE, ZERO } from '@/constants';
+import { DEFAULT_PAGE_SIZE, FALSE, TRUE, ZERO, authOptions } from '@/constants';
 import prisma from '@/libs/prisma';
 import type { Article, GetArticlesRequest } from '@/types';
 import {
@@ -24,7 +25,7 @@ export async function GET(req: Request) {
   const friendlyUrl = searchParams.get('friendlyUrl');
 
   const { tags = [] } = parse(
-    req.url?.split('?')[1] || '',
+    req.url?.split('?')[1] ?? '',
   ) as unknown as GetArticlesRequest;
 
   let take = DEFAULT_PAGE_SIZE;
@@ -89,7 +90,7 @@ export async function GET(req: Request) {
       tags: true,
     },
   });
-  const total = (await prisma.article.count({ where: condition })) || 0;
+  const total = (await prisma.article.count({ where: condition })) ?? 0;
 
   return NextResponse.json(
     createSuccessTotalResponse<Article[]>(articles, total),
