@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Metadata } from 'next';
+import { type Metadata } from 'next';
 
 import { GiscusComment } from '@/components/client';
 import { PageTitle } from '@/components/rsc';
