@@ -10,7 +10,7 @@ export default function GiscusComment() {
 
   useMount(() => {
     if (
-      localStorage.getItem('color-theme') === 'dark' ||
+      localStorage.getItem('color-theme') === 'dark' ??
       (!('color-theme' in localStorage) &&
         window.matchMedia('(prefers-color-scheme: dark)').matches)
     ) {
