@@ -12,7 +12,7 @@ export default function ToggleTheme({ className }: { className?: string }) {
   useMount(() => {
     // On page load or when changing themes, best to add inline in `head` to avoid FOUC
     if (
-      localStorage.getItem('color-theme') === 'dark' ||
+      localStorage.getItem('color-theme') === 'dark' ??
       (!('color-theme' in localStorage) &&
         window.matchMedia('(prefers-color-scheme: dark)').matches)
     ) {
