@@ -1,8 +1,8 @@
-import { ClassValue, clsx } from 'clsx';
+import { type ClassValue, clsx } from 'clsx';
 import { twMerge } from 'tailwind-merge';
 
 import { WEBSITE } from '@/constants';
-import { GeneralResponse, TotalResponse } from '@/types';
+import { type GeneralResponse, type TotalResponse } from '@/types';
 
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs));
